requirements.txt
streamlit
google-genai
python-dotenv
PyPDF2
faiss-cpu
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
